Updated Zumo with port expander and analog write
The update moved the "analog write" into a function that can be
copy-pasted. The same goes for the init of the port expander.

Wrote also a version that has a more "advanced" version of these
functions. The init function here only inits the ports used to outputs and
generates a mask for these ports. The "analog write" function will not
change the values on the ports that are not included in the output port
list.